requirements.txt
google-generativeai==0.8.5
google-api-core==2.17.1
protobuf==4.25.3
grpcio==1.60.1
python-dotenv==1.0.1
